requirements.txt
streamlit
msal
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyGithub